Split fixtures into 'stateless' and 'integration'
